move rel ns handling to xml_import
